core: special case AppManifest as baseArtifact
